Mov: reward mail support to separate module Del: MAIL_CLICK_SAFE_AREA, MAIL_COLLECT sufficient as substitute
