+++ module/mail/assets.py
@@ -1,0 +1,10 @@
+from module.base.button import Button
+from module.base.template import Template
+
+# This file was automatically generated by dev_tools/button_extract.py.
+# Don't modify it manually.
+
+MAIL_COLLECT = Button(area={'cn': (865, 583, 947, 601), 'en': (865, 583, 947, 601), 'jp': (865, 583, 947, 601), 'tw': (865, 583, 947, 601)}, color={'cn': (151, 180, 216), 'en': (151, 180, 216), 'jp': (151, 180, 216), 'tw': (151, 180, 216)}, button={'cn': (865, 583, 947, 601), 'en': (865, 583, 947, 601), 'jp': (865, 583, 947, 601), 'tw': (865, 583, 947, 601)}, file={'cn': './assets/cn/mail/MAIL_COLLECT.png', 'en': './assets/en/mail/MAIL_COLLECT.png', 'jp': './assets/jp/mail/MAIL_COLLECT.png', 'tw': './assets/tw/mail/MAIL_COLLECT.png'})
+MAIL_COLLECTED = Button(area={'cn': (835, 578, 975, 606), 'en': (835, 578, 975, 606), 'jp': (835, 578, 975, 606), 'tw': (835, 578, 975, 606)}, color={'cn': (54, 63, 71), 'en': (54, 63, 71), 'jp': (54, 63, 71), 'tw': (54, 63, 71)}, button={'cn': (835, 578, 975, 606), 'en': (835, 578, 975, 606), 'jp': (835, 578, 975, 606), 'tw': (835, 578, 975, 606)}, file={'cn': './assets/cn/mail/MAIL_COLLECTED.png', 'en': './assets/en/mail/MAIL_COLLECTED.png', 'jp': './assets/jp/mail/MAIL_COLLECTED.png', 'tw': './assets/tw/mail/MAIL_COLLECTED.png'})
+MAIL_DELETE = Button(area={'cn': (428, 567, 500, 584), 'en': (428, 567, 500, 584), 'jp': (428, 567, 500, 584), 'tw': (428, 567, 500, 584)}, color={'cn': (216, 173, 169), 'en': (216, 173, 169), 'jp': (216, 173, 169), 'tw': (216, 173, 169)}, button={'cn': (428, 567, 500, 584), 'en': (428, 567, 500, 584), 'jp': (428, 567, 500, 584), 'tw': (428, 567, 500, 584)}, file={'cn': './assets/cn/mail/MAIL_DELETE.png', 'en': './assets/en/mail/MAIL_DELETE.png', 'jp': './assets/jp/mail/MAIL_DELETE.png', 'tw': './assets/tw/mail/MAIL_DELETE.png'})
+MAIL_ENTER = Button(area={'cn': (1207, 393, 1253, 429), 'en': (1207, 393, 1253, 429), 'jp': (1207, 393, 1253, 429), 'tw': (1207, 393, 1253, 429)}, color={'cn': (109, 107, 95), 'en': (109, 107, 95), 'jp': (109, 107, 95), 'tw': (109, 107, 95)}, button={'cn': (1207, 393, 1253, 429), 'en': (1207, 393, 1253, 429), 'jp': (1207, 393, 1253, 429), 'tw': (1207, 393, 1253, 429)}, file={'cn': './assets/cn/mail/MAIL_ENTER.png', 'en': './assets/en/mail/MAIL_ENTER.png', 'jp': './assets/jp/mail/MAIL_ENTER.png', 'tw': './assets/tw/mail/MAIL_ENTER.png'})
