+++ module/reward/assets.py
@@ -6,11 +6,6 @@ from module.base.template import Template
 
 COIN = Button(area={'cn': (279, 48, 308, 73), 'en': (272, 53, 317, 73), 'jp': (279, 48, 308, 73), 'tw': (279, 48, 308, 73)}, color={'cn': (228, 175, 73), 'en': (221, 170, 75), 'jp': (228, 175, 73), 'tw': (228, 175, 73)}, button={'cn': (279, 48, 308, 73), 'en': (272, 53, 317, 73), 'jp': (279, 48, 308, 73), 'tw': (279, 48, 308, 73)}, file={'cn': './assets/cn/reward/COIN.png', 'en': './assets/en/reward/COIN.png', 'jp': './assets/jp/reward/COIN.png', 'tw': './assets/tw/reward/COIN.png'})
 EXP = Button(area={'cn': (449, 31, 477, 49), 'en': (455, 51, 490, 71), 'jp': (449, 31, 477, 49), 'tw': (449, 31, 477, 49)}, color={'cn': (145, 158, 176), 'en': (133, 133, 145), 'jp': (145, 158, 176), 'tw': (145, 158, 176)}, button={'cn': (449, 31, 477, 49), 'en': (455, 51, 490, 71), 'jp': (449, 31, 477, 49), 'tw': (449, 31, 477, 49)}, file={'cn': './assets/cn/reward/EXP.png', 'en': './assets/en/reward/EXP.png', 'jp': './assets/jp/reward/EXP.png', 'tw': './assets/tw/reward/EXP.png'})
-MAIL_CLICK_SAFE_AREA = Button(area={'cn': (396, 86, 461, 111), 'en': (396, 86, 461, 111), 'jp': (396, 86, 461, 111), 'tw': (396, 86, 461, 111)}, color={'cn': (176, 190, 221), 'en': (176, 190, 221), 'jp': (176, 190, 221), 'tw': (176, 190, 221)}, button={'cn': (396, 86, 461, 111), 'en': (396, 86, 461, 111), 'jp': (396, 86, 461, 111), 'tw': (396, 86, 461, 111)}, file={'cn': './assets/cn/reward/MAIL_CLICK_SAFE_AREA.png', 'en': './assets/en/reward/MAIL_CLICK_SAFE_AREA.png', 'jp': './assets/jp/reward/MAIL_CLICK_SAFE_AREA.png', 'tw': './assets/tw/reward/MAIL_CLICK_SAFE_AREA.png'})
-MAIL_COLLECT = Button(area={'cn': (865, 583, 947, 601), 'en': (865, 583, 947, 601), 'jp': (865, 583, 947, 601), 'tw': (865, 583, 947, 601)}, color={'cn': (151, 180, 216), 'en': (151, 180, 216), 'jp': (151, 180, 216), 'tw': (151, 180, 216)}, button={'cn': (865, 583, 947, 601), 'en': (865, 583, 947, 601), 'jp': (865, 583, 947, 601), 'tw': (865, 583, 947, 601)}, file={'cn': './assets/cn/reward/MAIL_COLLECT.png', 'en': './assets/en/reward/MAIL_COLLECT.png', 'jp': './assets/cn/reward/MAIL_COLLECT.png', 'tw': './assets/cn/reward/MAIL_COLLECT.png'})
-MAIL_COLLECTED = Button(area={'cn': (835, 578, 975, 606), 'en': (835, 578, 975, 606), 'jp': (835, 578, 975, 606), 'tw': (835, 578, 975, 606)}, color={'cn': (54, 63, 71), 'en': (54, 63, 71), 'jp': (54, 63, 71), 'tw': (54, 63, 71)}, button={'cn': (835, 578, 975, 606), 'en': (835, 578, 975, 606), 'jp': (835, 578, 975, 606), 'tw': (835, 578, 975, 606)}, file={'cn': './assets/cn/reward/MAIL_COLLECTED.png', 'en': './assets/en/reward/MAIL_COLLECTED.png', 'jp': './assets/cn/reward/MAIL_COLLECTED.png', 'tw': './assets/cn/reward/MAIL_COLLECTED.png'})
-MAIL_DELETE = Button(area={'cn': (428, 567, 500, 584), 'en': (428, 567, 500, 584), 'jp': (428, 567, 500, 584), 'tw': (428, 567, 500, 584)}, color={'cn': (216, 173, 169), 'en': (216, 173, 169), 'jp': (216, 173, 169), 'tw': (216, 173, 169)}, button={'cn': (428, 567, 500, 584), 'en': (428, 567, 500, 584), 'jp': (428, 567, 500, 584), 'tw': (428, 567, 500, 584)}, file={'cn': './assets/cn/reward/MAIL_DELETE.png', 'en': './assets/en/reward/MAIL_DELETE.png', 'jp': './assets/cn/reward/MAIL_DELETE.png', 'tw': './assets/cn/reward/MAIL_DELETE.png'})
-MAIL_ENTER = Button(area={'cn': (1207, 393, 1253, 429), 'en': (1207, 393, 1253, 429), 'jp': (1207, 393, 1253, 429), 'tw': (1207, 393, 1253, 429)}, color={'cn': (109, 107, 95), 'en': (109, 107, 95), 'jp': (109, 107, 95), 'tw': (109, 107, 95)}, button={'cn': (1207, 393, 1253, 429), 'en': (1207, 393, 1253, 429), 'jp': (1207, 393, 1253, 429), 'tw': (1207, 393, 1253, 429)}, file={'cn': './assets/cn/reward/MAIL_ENTER.png', 'en': './assets/en/reward/MAIL_ENTER.png', 'jp': './assets/jp/reward/MAIL_ENTER.png', 'tw': './assets/tw/reward/MAIL_ENTER.png'})
 MISSION_MULTI = Button(area={'cn': (1041, 8, 1101, 39), 'en': (1041, 8, 1101, 39), 'jp': (1041, 7, 1102, 36), 'tw': (1040, 6, 1102, 39)}, color={'cn': (226, 192, 142), 'en': (221, 179, 96), 'jp': (219, 178, 110), 'tw': (223, 184, 121)}, button={'cn': (1041, 8, 1101, 39), 'en': (1041, 8, 1101, 39), 'jp': (1041, 7, 1102, 36), 'tw': (1040, 6, 1102, 39)}, file={'cn': './assets/cn/reward/MISSION_MULTI.png', 'en': './assets/en/reward/MISSION_MULTI.png', 'jp': './assets/jp/reward/MISSION_MULTI.png', 'tw': './assets/tw/reward/MISSION_MULTI.png'})
 MISSION_NOTICE = Button(area={'cn': (940, 670, 945, 681), 'en': (940, 670, 945, 681), 'jp': (940, 670, 945, 681), 'tw': (940, 670, 945, 681)}, color={'cn': (183, 83, 66), 'en': (183, 83, 66), 'jp': (183, 83, 66), 'tw': (183, 83, 66)}, button={'cn': (940, 670, 945, 681), 'en': (940, 670, 945, 681), 'jp': (940, 670, 945, 681), 'tw': (940, 670, 945, 681)}, file={'cn': './assets/cn/reward/MISSION_NOTICE.png', 'en': './assets/en/reward/MISSION_NOTICE.png', 'jp': './assets/jp/reward/MISSION_NOTICE.png', 'tw': './assets/tw/reward/MISSION_NOTICE.png'})
 MISSION_SINGLE = Button(area={'cn': (1093, 118, 1179, 177), 'en': (1093, 118, 1179, 177), 'jp': (1102, 120, 1166, 149), 'tw': (1090, 115, 1181, 166)}, color={'cn': (115, 155, 218), 'en': (106, 147, 215), 'jp': (136, 176, 226), 'tw': (108, 149, 216)}, button={'cn': (1093, 118, 1179, 177), 'en': (1093, 118, 1179, 177), 'jp': (1102, 120, 1166, 149), 'tw': (1090, 115, 1181, 166)}, file={'cn': './assets/cn/reward/MISSION_SINGLE.png', 'en': './assets/en/reward/MISSION_SINGLE.png', 'jp': './assets/jp/reward/MISSION_SINGLE.png', 'tw': './assets/tw/reward/MISSION_SINGLE.png'})
